Split editor and runtime dependencies for composites, interactions and processors
