+++ Packages/com.unity.inputsystem/Editor/Actions/Composites.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: cfd47b21b6c654315bf28318c87beb98
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
