+++ Packages/com.unity.inputsystem/Editor/Actions/Composites/AxisCompositeEditor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 504f28d05365c4127b7216fd72484311
