+++ Packages/com.unity.inputsystem/Editor/Actions/Composites/Vector2CompositeEditor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 0eb07b1cf75d04357a7fbd95093f1fa9
