+++ Packages/com.unity.inputsystem/Editor/Actions/Composites/Vector3CompositeEditor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: d9d2c9fcdc7f646d083ee2360d5139bb
