+++ Packages/com.unity.inputsystem/Editor/Actions/Interactions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e4bf910aee5784821a3ebd59e4cdac79
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
