+++ Packages/com.unity.inputsystem/Editor/Actions/Interactions/HoldInteractionEditor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: fde8c91ad1cbc48e7a49f856f44724c6
