+++ Packages/com.unity.inputsystem/Editor/Actions/Interactions/MultiTapInteractionEditor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 991f8cb0b79a44a8687af2973bc223cc
