+++ Packages/com.unity.inputsystem/Editor/Actions/Interactions/PressInteractionEditor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: bb40b2b9ded5942e98788267f3adbc11
