+++ Packages/com.unity.inputsystem/Editor/Actions/Interactions/SlowTapInteractionEditor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 31783c2cca6e44f419396b9d094e2b31
