+++ Packages/com.unity.inputsystem/Editor/Actions/Interactions/TapInteractionEditor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 49b16a0da25fb4403a778600bd0278db
