+++ Packages/com.unity.inputsystem/Editor/Controls/Processors.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5e167cf5b25c44f35919e4082e725d71
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
