+++ Packages/com.unity.inputsystem/Editor/Controls/Processors/StickDeadzoneProcessorEditor.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: 58a253233c6804f58adf86cfee510a72
