+++ Packages/com.unity.inputsystem/Runtime/Actions/Interactions/SlowTapInteraction.cs
@@ -1,11 +1,6 @@
 using System;
 using System.ComponentModel;
 using UnityEngine.InputSystem.Controls;
-#if UNITY_EDITOR
-using UnityEngine.InputSystem.Editor;
-using UnityEngine.UIElements;
-#endif
-
 ////REVIEW: this is confusing when considered next to HoldInteraction; also it's confusingly named
 
 namespace UnityEngine.InputSystem.Interactions
@@ -66,37 +61,4 @@ namespace UnityEngine.InputSystem.Interactions
             m_SlowTapStartTime = 0.0;
         }
     }
-
-    #if UNITY_EDITOR
-    internal class SlowTapInteractionEditor : InputParameterEditor<SlowTapInteraction>
-    {
-        protected override void OnEnable()
-        {
-            m_DurationSetting.Initialize("Min Tap Duration",
-                "Minimum time (in seconds) that a control has to be held for it to register as a slow tap. If the control is released "
-                + "before this time, the slow tap is canceled.",
-                "Default Slow Tap Time",
-                () => target.duration, x => target.duration = x, () => InputSystem.settings.defaultSlowTapTime);
-            m_PressPointSetting.Initialize("Press Point",
-                "The amount of actuation a control requires before being considered pressed. If not set, default to "
-                + "'Default Button Press Point' in the global input settings.",
-                "Default Button Press Point",
-                () => target.pressPoint, v => target.pressPoint = v,
-                () => InputSystem.settings.defaultButtonPressPoint);
-        }
-
-        public override void OnGUI()
-        {
-        }
-
-        public override void OnDrawVisualElements(VisualElement root, Action onChangedCallback)
-        {
-            m_DurationSetting.OnDrawVisualElements(root, onChangedCallback);
-            m_PressPointSetting.OnDrawVisualElements(root, onChangedCallback);
-        }
-
-        private CustomOrDefaultSetting m_DurationSetting;
-        private CustomOrDefaultSetting m_PressPointSetting;
-    }
-    #endif
 }
