+++ Packages/com.unity.inputsystem/Runtime/Controls/Processors/StickDeadzoneProcessor.cs
@@ -1,10 +1,5 @@
 using System;
 
-#if UNITY_EDITOR
-using UnityEngine.InputSystem.Editor;
-using UnityEngine.UIElements;
-#endif
-
 ////REVIEW: rename to RadialDeadzone
 
 ////TODO: add different deadzone shapes and/or option to min/max X and Y separately
@@ -61,36 +56,4 @@ namespace UnityEngine.InputSystem.Processors
             return $"StickDeadzone(min={minOrDefault},max={maxOrDefault})";
         }
     }
-
-    #if UNITY_EDITOR
-    internal class StickDeadzoneProcessorEditor : InputParameterEditor<StickDeadzoneProcessor>
-    {
-        protected override void OnEnable()
-        {
-            m_MinSetting.Initialize("Min",
-                "Vector length  below which input values will be clamped. After clamping, vector lengths will be renormalized to [0..1] between min and max.",
-                "Default Deadzone Min",
-                () => target.min, v => target.min = v,
-                () => InputSystem.settings.defaultDeadzoneMin);
-            m_MaxSetting.Initialize("Max",
-                "Vector length above which input values will be clamped. After clamping, vector lengths will be renormalized to [0..1] between min and max.",
-                "Default Deadzone Max",
-                () => target.max, v => target.max = v,
-                () => InputSystem.settings.defaultDeadzoneMax);
-        }
-
-        public override void OnGUI()
-        {
-        }
-
-        public override void OnDrawVisualElements(VisualElement root, Action onChangedCallback)
-        {
-            m_MinSetting.OnDrawVisualElements(root, onChangedCallback);
-            m_MaxSetting.OnDrawVisualElements(root, onChangedCallback);
-        }
-
-        private CustomOrDefaultSetting m_MinSetting;
-        private CustomOrDefaultSetting m_MaxSetting;
-    }
-    #endif
 }
